requirements.txt
Django==1.8.6
django-model-utils==2.3.1
django-reversion==1.9.3
wheel==0.24.0
